allow plotting resolution bar in hkle resolution frame
